cy8c_tma_ts: Fix touchscreen not responding when resuming from sleep
Add a delay when the touchscreen resumes from sleep to ensure that it always has enough time to initialize. 50 was used as it was found to be near the minimum delay required for this kernel.

This bug can be accurately reproduced by pressing the power button many times in rapid succession.

Signed-off-by: Sultanxda <sultanxda@gmail.com>