pyramid: Add sysfs switch for basic color control
The sysfs switch allows switching between the old color configuration and the new one. Sysfs switch is located at /sys/module/mdp/parameters/mdp_gamma_cooler_colors. The cooler color profile is disabled by default. The sysfs switch can be written to on-the-go; turn the screen off and on to apply the new color profile.

Signed-off-by: Sultanxda <sultanxda@gmail.com>